requirements.txt
alembic==1.0.8
APScheduler==3.5.3
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.1
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.0
tzlocal==1.5.1
Werkzeug==0.14.1
WTForms==2.2.1
